Refactor dashboard component and update routing in App.jsx
